// src/privileges/enums/privilege-message.enum.ts
export enum PrivilegeMessage {
    NOT_FOUND = 'Privilege not found',
    ALREADY_EXISTS = 'Privilege already exists',
    CREATED = 'Privilege created successfully',
    UPDATED = 'Privilege updated successfully',
    DELETED = 'Privilege deleted successfully',
    GET_ALL_PRIVILEGES = 'Get all Privileges successfully',
    GET_PRIVILEGE = 'Get Privilege successfully',
}